Raise an error if attempting to use a Regexp pattern match on a database that doesn't support it
Before, Sequel used the PostgreSQL regexp format on most databases,
which would probably result in the database raising an error.  This
makes Sequel raise an exception while producing the SQL.

This adds Dataset#supports_regexp? for checking for support.  With
this change, integration tests can be added to check for correct
behavior, since they can be conditional on the supports_regexp?
method.