Don't require a :type when dropping constraints on SQLite
Previously, if a type was not specified, an error was raised.
However, other than primary key and foreign key constraints,
there are unique and check constraints, which can both be
handled similarly.

This allows you to drop check constraints on SQLite, though
it drops all check constraints instead of just the one listed.