Remove unique constraint in constraint validations table
This causes a problem on MySQL on certain charsets, see MySQL
bug 4541.  There isn't a need for a unique constraint, anyway.
The way the plugin works, it selects all columns from the table
which is going to do a sequential scan.  The only thing an index
would help is when dropping rows, and that is a fairly rare usage.
Considering that this table should never be very large, I don't
think any indexes are needed.